chore(preboot): rename "client" to "browser"
