--- modules/preboot/client.ts
@@ -1,1 +1,0 @@
-export * from './src/client/preboot_client';
