+++ modules/preboot/preboot.d.ts
@@ -1,2 +1,2 @@
-export function getClientCode(opts: any): any;
-export function getClientCodeStream(opts: any): any;
+export function getBrowserCode(opts: any): any;
+export function getBrowserCodeStream(opts: any): any;
