+++ modules/preboot/src/browser/buffer_manager.ts
