+++ modules/preboot/src/browser/freeze/freeze_with_spinner.ts
