+++ modules/preboot/test/preboot_karma.ts
@@ -1,12 +1,12 @@
 
 // this is the entry point for karma tests
-import './client/freeze/freeze_with_spinner_spec';
-import './client/listen/listen_by_attributes_spec';
-import './client/listen/listen_by_event_bindings_spec';
-import './client/listen/listen_by_selectors_spec';
-import './client/replay/replay_after_hydrate_spec';
-import './client/replay/replay_after_rerender_spec';
-import './client/buffer_manager_spec';
-import './client/dom_spec';
-import './client/event_manager_spec';
-import './client/log_spec';
+import './browser/freeze/freeze_with_spinner_spec';
+import './browser/listen/listen_by_attributes_spec';
+import './browser/listen/listen_by_event_bindings_spec';
+import './browser/listen/listen_by_selectors_spec';
+import './browser/replay/replay_after_hydrate_spec';
+import './browser/replay/replay_after_rerender_spec';
+import './browser/buffer_manager_spec';
+import './browser/dom_spec';
+import './browser/event_manager_spec';
+import './browser/log_spec';
