+++ tsconfig.json
@@ -17,39 +17,39 @@
     "tsd_typings/tsd.d.ts",
     "custom_typings/_custom.d.ts",
     "modules/index.ts",
-    "modules/preboot/client.ts",
+    "modules/preboot/browser.ts",
     "modules/preboot/server.ts",
-    "modules/preboot/src/client/buffer_manager.ts",
-    "modules/preboot/src/client/dom.ts",
-    "modules/preboot/src/client/event_manager.ts",
-    "modules/preboot/src/client/log.ts",
-    "modules/preboot/src/client/preboot_client.ts",
-    "modules/preboot/src/client/freeze/freeze_with_spinner.ts",
-    "modules/preboot/src/client/listen/listen_by_attributes.ts",
-    "modules/preboot/src/client/listen/listen_by_event_bindings.ts",
-    "modules/preboot/src/client/listen/listen_by_selectors.ts",
-    "modules/preboot/src/client/replay/replay_after_hydrate.ts",
-    "modules/preboot/src/client/replay/replay_after_rerender.ts",
-    "modules/preboot/src/server/client_code_generator.ts",
+    "modules/preboot/src/browser/buffer_manager.ts",
+    "modules/preboot/src/browser/dom.ts",
+    "modules/preboot/src/browser/event_manager.ts",
+    "modules/preboot/src/browser/log.ts",
+    "modules/preboot/src/browser/preboot_browser.ts",
+    "modules/preboot/src/browser/freeze/freeze_with_spinner.ts",
+    "modules/preboot/src/browser/listen/listen_by_attributes.ts",
+    "modules/preboot/src/browser/listen/listen_by_event_bindings.ts",
+    "modules/preboot/src/browser/listen/listen_by_selectors.ts",
+    "modules/preboot/src/browser/replay/replay_after_hydrate.ts",
+    "modules/preboot/src/browser/replay/replay_after_rerender.ts",
+    "modules/preboot/src/server/browser_code_generator.ts",
     "modules/preboot/src/server/normalize.ts",
     "modules/preboot/src/server/preboot_server.ts",
     "modules/preboot/src/server/presets.ts",
     "modules/preboot/src/server/utils.ts",
     "modules/preboot/test/preboot_karma.ts",
-    "modules/preboot/test/server/client_code_generator_spec.ts",
+    "modules/preboot/test/server/browser_code_generator_spec.ts",
     "modules/preboot/test/server/normalize_spec.ts",
     "modules/preboot/test/server/presets_spec.ts",
     "modules/preboot/test/server/utils_spec.ts",
-    "modules/preboot/test/client/buffer_manager_spec.ts",
-    "modules/preboot/test/client/dom_spec.ts",
-    "modules/preboot/test/client/event_manager_spec.ts",
-    "modules/preboot/test/client/log_spec.ts",
-    "modules/preboot/test/client/freeze/freeze_with_spinner_spec.ts",
-    "modules/preboot/test/client/listen/listen_by_attributes_spec.ts",
-    "modules/preboot/test/client/listen/listen_by_event_bindings_spec.ts",
-    "modules/preboot/test/client/listen/listen_by_selectors_spec.ts",
-    "modules/preboot/test/client/replay/replay_after_hydrate_spec.ts",
-    "modules/preboot/test/client/replay/replay_after_rerender_spec.ts",
+    "modules/preboot/test/browser/buffer_manager_spec.ts",
+    "modules/preboot/test/browser/dom_spec.ts",
+    "modules/preboot/test/browser/event_manager_spec.ts",
+    "modules/preboot/test/browser/log_spec.ts",
+    "modules/preboot/test/browser/freeze/freeze_with_spinner_spec.ts",
+    "modules/preboot/test/browser/listen/listen_by_attributes_spec.ts",
+    "modules/preboot/test/browser/listen/listen_by_event_bindings_spec.ts",
+    "modules/preboot/test/browser/listen/listen_by_selectors_spec.ts",
+    "modules/preboot/test/browser/replay/replay_after_hydrate_spec.ts",
+    "modules/preboot/test/browser/replay/replay_after_rerender_spec.ts",
     "modules/universal/server/index.ts",
     "modules/universal/server/server.ts",
     "modules/universal/server/src/platform/dom/server_dom_renderer.ts",
